machine: Split xml editing function out of libvirt-common
This file now contains functions for editing libvirt XMLs.
libvirt-common.js used to contain common provider function, mostly wrappers
around libvirt APIs. Now it's a file that should be merged with
libvirt-dbus.js and totally should not contain XML parsing functions.
